refactor(log): LogPage 파일 구조 개선 (#78)
