+++ app/(main)/log/constants.ts
@@ -1,1 +1,4 @@
 export const HEADER_TITLES = ['차량번호', '차종', '운행일수', '총운행거리', '차량현황']
+
+export const EXCEL_FILE_NAME = '차량운행기록.xlsx'
+export const EXCEL_SHEET_NAME = '차량운행기록'
