+++ components/domain/vehicle/ControlBar/index.tsx
@@ -9,11 +9,11 @@ import { useSearchSingleVehicle } from '@/hooks/useSearchSingleVehicle'
 
 import * as styles from './styles.css'
 
-interface SearchFieldProps {
-    hasButton: boolean
-    onExcelDownload?: () => void
+interface ControlBarProps {
+    onExcelButtonClick?: () => void
 }
-const SearchField = ({ hasButton, onExcelDownload }: SearchFieldProps) => {
+
+const ControlBar = ({ onExcelButtonClick }: ControlBarProps) => {
     const { modalMessage, isOpen, searchTerm, handleVehicleSearch, handleSearchChange, closeModal } =
         useSearchSingleVehicle()
 
@@ -27,22 +27,20 @@ const SearchField = ({ hasButton, onExcelDownload }: SearchFieldProps) => {
                     onSubmit={handleVehicleSearch}
                 />
             </div>
-            {hasButton && (
-                <div className={styles.buttonGroup}>
-                    <RoundButton size={'small'} color={'secondary'} onClick={onExcelDownload}>
-                        <div className={styles.button}>
-                            <Image src='/icons/green-excel-icon.svg' alt='excel' width={16} height={16} />
-                            엑셀
-                        </div>
-                    </RoundButton>
-                    <LinkButton href={'/log/register'}>
-                        <div className={styles.linkButton}>
-                            <Image src='/icons/white-add-icon.svg' alt='add' width={18} height={18} />
-                            등록
-                        </div>
-                    </LinkButton>
-                </div>
-            )}
+            <div className={styles.buttonGroup}>
+                <RoundButton size={'small'} color={'secondary'} onClick={onExcelButtonClick}>
+                    <div className={styles.button}>
+                        <Image src='/icons/green-excel-icon.svg' alt='excel' width={16} height={16} />
+                        엑셀
+                    </div>
+                </RoundButton>
+                <LinkButton href={'/log/register'}>
+                    <div className={styles.linkButton}>
+                        <Image src='/icons/white-add-icon.svg' alt='add' width={18} height={18} />
+                        등록
+                    </div>
+                </LinkButton>
+            </div>
             <Modal
                 isOpen={isOpen}
                 message={modalMessage as ModalMessageType}
@@ -53,4 +51,4 @@ const SearchField = ({ hasButton, onExcelDownload }: SearchFieldProps) => {
     )
 }
 
-export default SearchField
+export default ControlBar
